refactor RequestForm. add styling
