+++ frontend/src/components/RequestForm/RequestForm.css
@@ -1,0 +1,14 @@
+.form-content {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.form-content-container {
+    display: flex;
+    flex-direction: column;
+}
+
+.form-group input[type=text] {
+    width: 100%;
+}
