+++ frontend/src/components/TextAreaInput/TextAreaInput.css
@@ -1,5 +1,6 @@
 .form-group textarea {
     height: 11rem;
+    width: 100%;
     padding: 0.75rem 1rem;
     border-radius: 0.75rem;
     border: 1px solid #e7e6e4;
