+++ frontend/src/components/TextInput/TextInput.css
@@ -12,4 +12,4 @@
         border-color: #2f6fed;
         outline: none;
     }
-} 
+}
